feat: Set-up algorithm files
Set-up in a way to maximise flexibility. A function should be able
to be simply added to suggest_pub.py, and the .env variable
updated to select this function. pub-service.py does the unpacking
of my database objects, and is in a new directory as it is a seperate
idea to my CRUD endpoints and allows me to update the unpacking
of object without affecting the algo. Finally, I made the algos a module
as I consider it be be its own entity and this means I can import it
freely across the rest of the code base and write tests etc more
easily and makes it more modular.